+++ sandbox/urls.py
@@ -2,14 +2,16 @@
 URL Configuration for sandbox
 """
 from django.conf import settings
-from django.conf.urls.i18n import i18n_patterns
 from django.conf.urls.static import static
 from django.contrib import admin
 from django.contrib.staticfiles.urls import staticfiles_urlpatterns
-from django.urls import include, path
+from django.urls import path
+
+from django.views.generic import TemplateView
 
 
 urlpatterns = [
+    path("", TemplateView.as_view(template_name="home.html")),
     path("admin/", admin.site.urls),
     # path("", include("diskette.urls")),
 ]