+++ tests/conftest.py
@@ -27,7 +27,6 @@ class FixturesSettingsTestMixin(object):
         self.application_path = Path(
             diskette.__file__
         ).parents[0].resolve()
-        self.application_urlpath = ""
 
         self.package_path = self.application_path.parent